Make opening & closing idempotent and add events
Very similar to `levelup` but more precise. If `open()` and
`close()` are called repeatedly (while the previous call has not
yet completed) the last call dictates the final status. Callbacks
are not called until any pending state changes are done, meaning
that the status is not 'opening' or 'closing'. Same for events.

For example, in a sequence of calls like `open(); close(); open()`
the final status will be 'open', only the second call will error,
only an 'open' event is emitted, and all callbacks will see that
status is 'open'. The callbacks are called in the order that the
`open()` or `close()` calls were made.

In addition, unlike on `levelup`, it is safe to call `open()` while
status is 'closing'. It will wait for closing to complete and then
reopen.

We should now have complete safety, including in `leveldown` because
the native code there delays `close()` if any operations are in
flight. In other words, the JavaScript side in `abstract-leveldown`
prevents new operations before opening the db, and the C++ side in
`leveldown` prevents closing the db before operations completed.

Ref https://github.com/Level/leveldown/issues/8
Ref https://github.com/Level/community/issues/58